+++ src/freenet/node/fcp/ProbeStoreSize.java
@@ -6,9 +6,9 @@ package freenet.node.fcp;
 public class ProbeStoreSize extends FCPResponse {
 	/**
 	 * @param fcpIdentifier FCP-level identifier for pairing requests and responses
-	 * @param storeSize reported endpoint store size in GiB.
+	 * @param storeSize reported endpoint store size in GiB multiplied by Gaussian noise.
 	 */
-	public ProbeStoreSize(String fcpIdentifier, long storeSize) {
+	public ProbeStoreSize(String fcpIdentifier, float storeSize) {
 		super(fcpIdentifier);
 		fs.put(STORE_SIZE, storeSize);
 	}
